optimize css viewmodel structure
