+++ NGOPlatformWeb/Models/ViewModels/Purchase/PurchaseViewModels.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace NGOPlatformWeb.Models.ViewModels
+namespace NGOPlatformWeb.Models.ViewModels.Purchase
 {
     /// <summary>
     /// 物資認購的基本ViewModel
