+++ NGOPlatformWeb/Views/Auth/Login.cshtml
@@ -9,7 +9,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>登入 - NGO 平台</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" />
-    <link rel="stylesheet" href="@Url.Content("~/css/login.css")" />
+    <link rel="stylesheet" href="@Url.Content("~/css/auth/login.css")" />
     <script src="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/js/all.min.js"></script>
 </head>
