+++ NGOPlatformWeb/Views/Case/CaseProfile.cshtml
@@ -1,11 +1,11 @@
-﻿@model NGOPlatformWeb.Models.ViewModels.CaseProfileViewModel
+﻿@model NGOPlatformWeb.Models.ViewModels.Profile.CaseProfileViewModel
 @{
     ViewData["Title"] = "個案個人資料";
 }
 
 @section Styles {
-    <link rel="stylesheet" href="~/css/profile-shared.css" />
-    <link rel="stylesheet" href="~/css/profile-components.css" />
+    <link rel="stylesheet" href="~/css/profile/profile-shared.css" />
+    <link rel="stylesheet" href="~/css/profile/profile-components.css" />
 }
 
 <div class="container py-5 case-page">
