+++ NGOPlatformWeb/Views/Case/Registrations.cshtml
@@ -1,12 +1,12 @@
 <!-- 個案活動參與紀錄頁面 - 顯示個案所有活動參與歷史 -->
-@model NGOPlatformWeb.Models.ViewModels.CaseActivityRegistrationsViewModel
+@model NGOPlatformWeb.Models.ViewModels.ActivityRegistrations.CaseActivityRegistrationsViewModel
 @{
     ViewData["Title"] = "活動參與紀錄";
 }
 
 @section Styles {
-    <link rel="stylesheet" href="~/css/profile-shared.css" />
-    <link rel="stylesheet" href="~/css/activity-records.css" />
+    <link rel="stylesheet" href="~/css/profile/profile-shared.css" />
+    <link rel="stylesheet" href="~/css/activities/activity-records.css" />
 }
 
 <div class="container py-5 case-page">
