+++ NGOPlatformWeb/Views/Home/Index.cshtml
@@ -4,7 +4,7 @@
 }
 
 @section Styles {
-    <link rel="stylesheet" href="~/css/homepage-new.css" />
+    <link rel="stylesheet" href="~/css/shared/homepage-new.css" />
 }
 
 <!-- 絲滑輪播圖區域 -->
