+++ NGOPlatformWeb/Views/Purchase/Index.cshtml
@@ -4,7 +4,7 @@
 
 @Html.AntiForgeryToken()
 
-<link rel="stylesheet" href="~/css/purchase.css" asp-append-version="true" />
+<link rel="stylesheet" href="~/css/purchase/purchase.css" asp-append-version="true" />
 
 
 @if (ViewBag.Error != null)
