+++ NGOPlatformWeb/Views/Purchase/Payment.cshtml
@@ -1,10 +1,10 @@
 ﻿<!-- 付款確認頁面 - 顯示付款資訊並處理付款流程 -->
-@model PaymentViewModel
+@model NGOPlatformWeb.Models.ViewModels.Purchase.PaymentViewModel
 @{
     ViewData["Title"] = "確認付款";
 }
 
-<link rel="stylesheet" href="~/css/payment.css" asp-append-version="true" />
+<link rel="stylesheet" href="~/css/purchase/payment.css" asp-append-version="true" />
 
 <div class="payment-page">
     <div class="container">
