+++ NGOPlatformWeb/Views/Purchase/Success.cshtml
@@ -1,9 +1,9 @@
-﻿@model OrderResultViewModel
+﻿@model NGOPlatformWeb.Models.ViewModels.Purchase.OrderResultViewModel
 @{
     ViewData["Title"] = "捐贈成功";
 }
 
-<link rel="stylesheet" href="~/css/payment.css" asp-append-version="true" />
+<link rel="stylesheet" href="~/css/purchase/payment.css" asp-append-version="true" />
 
 <div class="success-page">
     <div class="container">
