+++ NGOPlatformWeb/Views/User/EditProfile.cshtml
@@ -1,4 +1,4 @@
-﻿@model NGOPlatformWeb.Models.ViewModels.UserEditViewModel
+﻿@model NGOPlatformWeb.Models.ViewModels.Profile.UserEditViewModel
 @{
     ViewData["Title"] = "編輯個人資料";
 }
