+++ NGOPlatformWeb/Views/User/PurchaseRecords.cshtml
@@ -1,12 +1,12 @@
 <!-- 使用者認購紀錄頁面 - 顯示所有物資捐贈歷史 -->
-@model UserPurchaseRecordsViewModel
+@model NGOPlatformWeb.Models.ViewModels.Purchase.UserPurchaseRecordsViewModel
 @{
     ViewData["Title"] = "我的認購紀錄";
 }
 
 @section Styles {
-    <link rel="stylesheet" href="~/css/profile-shared.css" />
-    <link rel="stylesheet" href="~/css/purchase-records.css" />
+    <link rel="stylesheet" href="~/css/profile/profile-shared.css" />
+    <link rel="stylesheet" href="~/css/purchase/purchase-records.css" />
 }
 
 <div class="container py-5">
