+++ NGOPlatformWeb/Views/User/Registrations.cshtml
@@ -1,12 +1,12 @@
 <!-- 使用者活動報名紀錄頁面 - 顯示所有活動參與歷史 -->
-@model NGOPlatformWeb.Models.ViewModels.UserActivityRegistrationsViewModel
+@model NGOPlatformWeb.Models.ViewModels.ActivityRegistrations.UserActivityRegistrationsViewModel
 @{
     ViewData["Title"] = "活動報名紀錄";
 }
 
 @section Styles {
-    <link rel="stylesheet" href="~/css/profile-shared.css" />
-    <link rel="stylesheet" href="~/css/activity-records.css" />
+    <link rel="stylesheet" href="~/css/profile/profile-shared.css" />
+    <link rel="stylesheet" href="~/css/activities/activity-records.css" />
 }
 
 <div class="container py-5">
