+++ NGOPlatformWeb/Views/User/UserProfile.cshtml
@@ -1,12 +1,12 @@
 ﻿<!-- 一般使用者個人資料頁面 - 顯示基本資料、活動統計、認購統計 -->
-@model NGOPlatformWeb.Models.ViewModels.UserProfileViewModel
+@model NGOPlatformWeb.Models.ViewModels.Profile.UserProfileViewModel
 @{
     ViewData["Title"] = "個人資料";
 }
 
 @section Styles {
-    <link rel="stylesheet" href="~/css/profile-shared.css" />
-    <link rel="stylesheet" href="~/css/profile-components.css" />
+    <link rel="stylesheet" href="~/css/profile/profile-shared.css" />
+    <link rel="stylesheet" href="~/css/profile/profile-components.css" />
 }
 
 <div class="container py-5">
