+++ NGOPlatformWeb/wwwroot/css/activities/activity-records.css
