+++ NGOPlatformWeb/wwwroot/css/purchase/purchase.css
